Add info command to the python site information
